weather_web_application/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
haversine==2.8.0
numpy==1.24.2
pandas==1.5.2
pmdarima==2.0.3
Requests==2.31.0
scikit_learn==1.2.2
scipy==1.11.1
statsmodels==0.13.5
torch==2.0.0
tqdm==4.65.0
